LayoutTests/http/tests/security/xss-DENIED-window-open-javascript-url-with-spaces-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8080/security/resources/innocent-victim.html from frame with URL http://127.0.0.1:8000/security/xss-DENIED-window-open-javascript-url-with-spaces.html. Domains, protocols and ports must match.

This test passes if there is no alert dialog.
